// src/app/models/giaoviec/optionscv.model.ts
export class OptionsCV {
    Id: string;
    MyWorkId: string;
    p: number;
    pz: number;
    startPage?: number;
    endPage?: number;
    totalItem?: number;
    constructor() {
        this.p = 1;
        this.pz = 100;
        this.startPage = 0;
        this.pz = 100;
        this.totalItem = 100;
    }
}